recommendation-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
py2neo